rag_backend_fastapi/requirements.txt
# FastAPI và server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.4.0

# Vector Database
chromadb==0.4.22

# Embeddings
sentence-transformers>=2.3.0

# LLM
google-generativeai==0.3.2

# File processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
numpy>=1.26.0

